requirements.txt

dotenv==0.9.9

Flask==3.1.0
flask-cors==5.0.1


gunicorn==23.0.0

google-api-python-client==2.184.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2